% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalBF.R
\name{boundingOperation}
\alias{boundingOperation}
\title{Bounding Operation}
\usage{
boundingOperation(v, left, right, m, n)
}
\arguments{
\item{v}{a a numeric value giving an ordering permutation of 1 to m+n.}

\item{left}{a numeric value giving the number of witnesses left.}

\item{right}{a numeric value giving the number of witnesses right.}

\item{m}{a numeric value, the number of observations left.}

\item{n}{a numeric value, the number of observations right.}
}
\value{
a cumulative counting function represented as a numeric vector.
}
\description{
Bounding Operation
}
