% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWit.R
\name{dWit}
\alias{dWit}
\title{Lower-bound estimate of the number of distributional witness & total variation distance
for mixture models}
\usage{
dWit(t, rho, s, alpha = 0.05, estimator.type = "basic",
  tv.grid = seq(from = 0, to = 1, by = 0.01), direction = rep("left",
  length(s)))
}
\arguments{
\item{t}{a numeric value corresponding to an ordering of the observations. For two-sample
test times could be 0-1 numeric values values.}

\item{rho}{a numeric value giving an ordering function. This could be
a classifier, a regressor, a witness function from an MMD kernel or anything else.}

\item{s}{a numeric value giving split points on t.}

\item{alpha}{the level of the type 1 error control.}

\item{estimator.type}{either "basic", "tv-search" or "wit-search"}

\item{tv.grid}{a grid of values for the search.}

\item{direction}{which witness to estimatw.}
}
\description{
Lower-bound estimate of the number of distributional witness & total variation distance
for mixture models
}
