% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWitMatrix.R
\name{dWitMatrix}
\alias{dWitMatrix}
\title{Pairwise Total Variation Distance Lower Bound Matrix in Multi-Class Settings}
\usage{
dWitMatrix(labels, ordering.array, alpha = 0.05,
  computation.type = "non-optimized", seed = 0, ...)
}
\arguments{
\item{labels}{a numeric vector value. The labels of the classes, should be encoded in [0,nclass-1].}

\item{ordering.array}{a numeric array of size (nclass, nclass, nobs) such that the value (i,j,k) represents a propensity of beeing of class j instead of i for observation k.}

\item{alpha}{a numeric value. The type-I error level.}

\item{computation.type}{a character value. For the moment only "non-optimized" (default) available.}

\item{...}{additionnal parameters to be passed to the dWit function.}
}
\value{
a numeric matrix of size (nclass, nclass) giving the matrix of pairwise total variation lower bounds.
}
\description{
Pairwise Total Variation Distance Lower Bound Matrix in Multi-Class Settings
}
\examples{
 # iris example
 require(dWit)
 require(ranger)

 # training a multi-class classifier on iris and getting tv lower bounds between classes
 data("iris")

 ind.train <- sample(1:nrow(iris), size = nrow(iris)/2, replace = FALSE)

 rf <- ranger(Species~., data = iris[ind.train, ], probability = TRUE)
 preds <- predict(rf, iris[-ind.train,])$predictions

 # creating the ordering array based on prediction differences
 ar <- array(dim = c(3, 3, nrow(preds)))
 for (i in 1:3) {
   for (j in 1:3) {
    ar[i,j,] <- preds[,j] - preds[,i]
   }
 }

 # encoding the class response
 y <- factor(iris$Species)
 levels(y) <- c(0,1,2)

 # getting the lower bound matrix
 tvhat.iris <- dWit(labels = y[-ind.train], ordering.array = ar)
 tvhat.iris

}
