% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWitSampling.R
\name{dWitSampling}
\alias{dWitSampling}
\title{Distributional Witness Sampling}
\usage{
dWitSampling(x, f, g, seed = NULL)
}
\arguments{
\item{x}{a numeric matrix of observations as rows.}

\item{f}{a function giving the density left.}

\item{g}{a function giving the density right.}

\item{seed}{a numeric value for reproducibility.}
}
\value{
a list with values:
  \itemize{-dwit.f, dwit.g the witness indicators left and right respectively.
           -pdwit.f, pdwit.g the corresponding probabilities.}
}
\description{
Distributional Witness Sampling
}
