% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwlb.R
\name{dwlb}
\alias{dwlb}
\title{Lower-bound estimate of the number of distributional witness & total variation distance
for mixture models or two-sample tests}
\usage{
dwlb(times, preds, s, verbose.plot = FALSE, alpha = 0.05)
}
\arguments{
\item{times}{a numeric value corresponding to the times of the observations. For two-sample
test times could be 0-1 values.}

\item{preds}{a numeric value giving the ordering of the observations. This could come either from
a classification model or a regression model.}

\item{s}{the time value at which we consider a split and test for difference in distribution between the
the mixture sitting at the left of s and the one at the right of s.}

\item{verbose.plot}{a logical value indicating if we want some summary plots.}

\item{alpha}{the level of the type 1 error control.}
}
\description{
Lower-bound estimate of the number of distributional witness & total variation distance
for mixture models or two-sample tests
}
