% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalBF.R
\name{empiricalBF}
\alias{empiricalBF}
\title{Empirical Bounding Functions}
\usage{
empiricalBF(tv.seq, nsim = 1000, m = 100, n = 100, alpha = 0.05)
}
\arguments{
\item{tv.seq}{a vector of total variation values between 0 and 1.}

\item{nsim}{a numeric value giving the number of repetitions.}

\item{m}{a numeric value, the number of observations left.}

\item{n}{a numeric value, the number of observations right.}

\item{alpha}{a numeric value giving the type-I error level.}
}
\value{
a list of empirical bounding functions indexed by the tv.seq (in the respective order).
}
\description{
Empirical Bounding Functions
}
