% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTvLbDistanceMatrix.R
\name{getTvLbDistanceMatrix}
\alias{getTvLbDistanceMatrix}
\title{Get Distance}
\usage{
getTvLbDistanceMatrix(labels, ordering.array, alpha = 0.05,
  computation.type = "non-optimized")
}
\arguments{
\item{ordering.array}{array of size (nclass, nclass, nobs) such that the value (i,j,k) represents a propensity of beeing of class j instead of i for observation k.}

\item{alpha}{type I error control}
}
\description{
Get Distance
}
