% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertBinMeanTest.R
\name{invertBinMeanTest}
\alias{invertBinMeanTest}
\title{Inverting a Binomial Test}
\usage{
invertBinMeanTest(n.success, n.trial, alpha, rule.of.three = TRUE)
}
\arguments{
\item{n.success}{number of successes.}

\item{n.trial}{number of trials.}

\item{alpha}{the level.}

\item{rule.of.three}{should we apply the rule of three?}
}
\description{
Inverting a Binomial Test
}
