% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSampleTvLb.R
\name{twoSampleTvLb}
\alias{twoSampleTvLb}
\title{Two Sample Total Variance Lower Bound Estimate}
\usage{
twoSampleTvLb(labels, preds, alpha = 0.05)
}
\arguments{
\item{labels}{a factor value with levels 0 and 1 encoding the membership to the two samples.}

\item{preds}{a numeric value between 0 and 1 giving the probability predictions for class 1.}

\item{alpha}{a value giving the type 1 error control.}
}
\value{
the lower bound estimate
}
\description{
Two Sample Total Variance Lower Bound Estimate
}
